requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP & API
requests>=2.31.0
sodapy>=1.8.0
urllib3>=2.0.0

# Cron
python-crontab>=3.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
